fix: reuse SModel index for diagram pieces, recalculate LOD when new diagram pieces arrive. Moved diagram piece code together in a single folder
